% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aic.R
\name{compute_aic}
\alias{compute_aic}
\title{Function that just the AIC for the model data fits}
\usage{
compute_aic(observations, model, n_params)
}
\arguments{
\item{observations}{Vector of measured values}

\item{model}{Vector of model values}

\item{n_params}{number of parameters in model}
}
\value{
The computed AIC
}
\description{
\code{compute_aic} computes the AIC values
}
