% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_model_respiration.R
\name{compute_model_respiration}
\alias{compute_model_respiration}
\title{Compute modeled soil respiration from parameters and data}
\usage{
compute_model_respiration(input_data, input_params, input_model_exp)
}
\arguments{
\item{input_data}{Data used to compute respiration}

\item{input_model_exp}{Model expression used to compute respiration}

\item{input_paramms}{Parameter values used in expression}
}
\description{

}
