% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_model_respiration_components.R
\name{compute_model_respiration_components}
\alias{compute_model_respiration_components}
\title{Compute modeled soil respiration components from parameters and data}
\usage{
compute_model_respiration_components(
  input_data,
  input_params,
  input_ra_exp,
  input_rh_exp
)
}
\arguments{
\item{input_data}{Data used to compute respiration}

\item{input_ra_exp}{Model expression used to compute RA}

\item{input_rh_exp}{Model expression used to compute RH}

\item{input_paramms}{Parameter values used in expression}
}
\description{
\code{compute_model_respiration_components} Determines modeled soil respiration given data, parameters, and an expression
}
