% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_expressions.R
\docType{data}
\name{model_expressions}
\alias{model_expressions}
\title{Model expressions to compute the heterotrophic and autotrophic respiration}
\format{
A data frame with 5 rows, 4 columns.
}
\source{
Expressions processed and exported in data-process/06-model-expression-prepare.R
}
\usage{
data(model_expressions)
}
\description{
\itemize{
  \item model. One of null, microbe, quality, microbe-mult, or quality-mult
  \item incubation_expressions. The formula needed to compute soil respiration from incubation data
  \item field_expressions  The formula needed to compute soil respiration from field data
  \item rev_field_expressions. A modified formula for the Field Linear and Incubation Field Linear approaches
}
}
\keyword{datasets}
