% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimate.R
\name{parameter_estimate}
\alias{parameter_estimate}
\title{Parameter estimate using the Nash equivalent of LM algorithm.}
\usage{
parameter_estimate(parameters, input_data, input_expression)
}
\arguments{
\item{parameters}{a vector of parameters we use to estimate data}

\item{input_data}{data we use for the model}

\item{input_expression}{the point expression we use for the model}
}
\description{
\code{parameter_estimate} Determines a parameter estimate from input data.  I had to modify if from nlxb so it doesn't print out the progress messages.
}
