% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiration_expressions.R
\docType{data}
\name{respiration_expressions}
\alias{respiration_expressions}
\title{Respiration expressions to compute the heterotrophic and autotrophic respiration.  For convenience - this is the right hand side of RS = RA + RH}
\format{
A data frame with 5 rows, 3 columns.
}
\source{
Expressions processed and exported in data-process/06-model-expression-prepare.R
}
\usage{
data(respiration_expressions)
}
\description{
\itemize{
  \item model. One of null, microbe, quality, microbe-mult, or quality-mult
  \item RA_expression. The formula needed to compute autotrophic respiration
  \item RH_expression  The formula needed to heterotrophic respiration
}
}
\keyword{datasets}
