% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_filter.R
\name{rss_filter}
\alias{rss_filter}
\title{Filter the rss to the first third of the rss confidence interval, removing any duplicates in the parameters}
\usage{
rss_filter(input_results)
}
\arguments{
\item{input_results}{A nested data frame}
}
\value{
A filtered nested data frame
}
\description{
\code{rss_filter} Computes the summary distribution for the residual sum of squares and then filters out the unique values.  Has to work on a nested data frame
}
\examples{

# To be filled in later
}
