% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_median_parameters.R
\name{summarize_median_parameters}
\alias{summarize_median_parameters}
\title{From a listing of parameter estimates, filter on the rss between the 25th - 75h percentile, compute the median parameter estimates and the associated soil respiration}
\usage{
summarize_median_parameters(input_results, approach_name)
}
\arguments{
\item{input_results}{a nested (by Year, model, depth) vector of parameter estimates}

\item{approach_name}{name of the approach for labeling}
}
\description{
\code{summarize_median_parameters} returns the median parameters from a nested parameter estimations
}
