% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_median_soil_respiration.R
\name{summarize_median_soil_respiration}
\alias{summarize_median_soil_respiration}
\title{From a listing of parameter estimates, filter on the rss between the 25th - 75h percentile, compute the median parameter estimates and the associated soil respiration}
\usage{
summarize_median_soil_respiration(
  input_results,
  input_expressions,
  input_data,
  approach_name
)
}
\arguments{
\item{input_results}{a nested (by Year, model, depth) vector of parameter estimates}

\item{input_expressions}{expressions we use to compute the model}

\item{approach_name}{name of the approach for labeling}

\item{input_expression}{data used to compute respiration}
}
\description{
\code{summarize_median_soil_respiration} summarizes the soil respiration from the median parameters
}
