% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_compute.R
\name{taylor_compute}
\alias{taylor_compute}
\title{Function that just computes Taylor stats for the model and the data}
\usage{
taylor_compute(modeled, measured)
}
\arguments{
\item{modeled}{Vector of modeled values}

\item{measured}{Vector of measured values}
}
\value{
A data frame of results: sd_meas, sd_model (standard deviations of measured and modeled), r = correlation coefficient, centered root mean squared, and x and y coordinates for a taylor diagram)
}
\description{
\code{taylor_compute} computes all necessary information to produce a Taylor diagram
}
