% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_canada.R
\name{theme_canada}
\alias{theme_canada}
\title{Define a custom theme for making ggplots}
\usage{
theme_canada()
}
\description{
\code{theme_canada} Sets up defaults for ggplots so I don't need to keep doing these over and over.
}
